style 과거 했던 프로젝트 이동

+++ src/main/step2.lotto/automatic/domain/LottoMatch.java
@@ -37,10 +37,11 @@ public class LottoMatch {
     }
 
     public double getlottoProfitAmount() {
-        double returnValue = 0;
-        int totalPrice = LottoProfitAmount.runLottoProfitAmount();
+        double returnValue = 0.0;
+        double totalPrice = LottoProfitAmount.runLottoProfitAmount();
 
         returnValue = totalPrice / lottoPurchaseAmount;
+
         return returnValue;
     }
